Implement the waitForSpineWasmInstantiation interface in a separate file.
