+++ cocos/spine/lib/instantiated.ts
@@ -25,7 +25,7 @@
 import { BUILD, JSB, LOAD_SPINE_MANUALLY } from 'internal:constants';
 import { game } from '../../game';
 import './spine-define'; // Make spine._overrideSpineDefine take effect.
-import { waitForSpineWasmInstantiation } from './spine-version';
+import { waitForSpineWasmInstantiation } from './spine-instantiate';
 
 if (!JSB && (!BUILD || !LOAD_SPINE_MANUALLY)) {
     game.onPostInfrastructureInitDelegate.add(waitForSpineWasmInstantiation);
