+++ cocos/spine/skeleton.ts
@@ -45,8 +45,8 @@ import SkeletonCache, { AnimationCache, AnimationFrame, SkeletonCacheItemInfo } 
 import { TrackEntryListeners } from './track-entry-listeners';
 import { setPropertyEnumType } from '../core/internal-index';
 import { RenderData } from '../2d/renderer/render-data';
+import { SPINE_VERSION } from './lib/spine-version';
 
-const SPINE_VERSION = (spine as any).SPINE_VERSION;// defined in spine-version.ts
 const CachedFrameTime = 1 / 60;
 
 type TrackListener = (x: spine.TrackEntry) => void;
